sim-ctrl: support requeuing games with invalid reports
If the game report was not properly stored, the job would be considered
as completed but the game has no score.
Add a flag that allows to detect this situation and requeue the invalid
jobs again.

Eventually, run-game should make sure that it stores a valid game
report. For now, this workaround suffices for the few games where this
occurs.