sim-ctrl: more stable namespace determination
The namespace was so far only determined by the index in the ordered
list of job pod names. However, if a job had to be restarted, the order
could change and it would pick the wrong namespace.

Fix this by labeling job pods with the namespace they use to run the
simulation. Then, during querying, only pick namespace names out of the
set of available names. Still do this based on the ordered set names.
While this does not cover some situations, these are sufficiently
unlikely in our case that we accept this risk rather than going in full
using, e.g., etcd or MongoDB for synchronization.