sim-ctrl: gracefully handle TERM signal
When receiving the TERM signal, stop the simulation and requeue the
running game. We do this by raising an exception in the signal. This
will trigger PodController.monitor_pods to exit and requeue. Again, this
is not the best solution, but one easy to introduce and covering
virtually all of the tricky situations.