sim-ctrl: support permutation shuffling in create-tournament
Sometimes it may be desirable to shuffle each permutation for a
tournament. For example, intermediate results are less biased due to
some teams already having played a number of consecutive games. It may
also appear more fair and reduce the potential impact of system services
consuming some CPU time on a node for a short while for a single team.