sim-ctrl: reset manifests, optional failure time on requeue
When requeing an item, make it possible to do so without marking it as a
failure. This can come handy when a reconfiguration requires a game to
be run again, not making this an actual failure.

Also unset the generated manifests on the job. This is a little bit of
mixing concerns here. Might go about this differently another time.