sim-ctrl: allow for updating job parameters for existing jobs
If the game template document is modified, e.g., to fix an error, so far
it always required to create new jobs. The new tool allows to update
existing jobs by re-generating the parameters (based on the same input
data).

The procedure generates and stores a unified diff along with the
modification date into the database to make this traceable.